Move DatabaseAdapter, Reservation, and InventoryItem to backend.database
